+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All features and notable changes to this project will be documented in this file
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [0.0.8] - 2023-03-04
+
+### Added
+- (tcp) reacting to BYE message from server
+- (tcp) handling interrupt signal from user while waiting for auth verification
+- (tcp) added (5sec) timeout to waiting for auth verification
+
 ## [0.0.7] - 2023-03-04
 
 ### Added